regtests: split up base types and actual test implementation
The single python file started becoming pretty long so lets at least
split up base infrastructure and test case implementations.
